refactor(gui): extract settings state logic
